backend/requirements.txt
fastapi
huggingface_hub
pillow
python-dotenv
python-multipart
requests
supabase
uvicorn
